requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
seaborn>=0.12
scikit-learn>=1.3
streamlit>=1.30
oracledb>=1.4
python-dotenv>=1.0
